Split redeem component from sell/positions component
Show sidebar for disconnected users when a market is resolved
