+++ app/src/Sidebar/components/Sell/index.js
@@ -2,16 +2,13 @@ import React, { useState, useEffect, useCallback } from "react";
 import PropTypes from "prop-types";
 import Web3 from "web3";
 import { calcPositionGroups } from "utils/position-groups";
-import { getPositionId, combineCollectionIds } from "utils/getIdsUtil";
 import { getMarketProbabilities } from "utils/probabilities";
 
-const { BN, toBN, sha3 } = Web3.utils;
+const { BN } = Web3.utils;
 
-import getConditionalTokensRepo from "repositories/ConditionalTokensRepo";
 import getConditionalTokensService from "services/ConditionalTokensService";
 import Sell from "./SellForm";
 import Positions from "./Positions";
-let conditionalTokensRepo;
 let conditionalTokensService;
 
 // let warnedAboutIds = {};
@@ -19,7 +16,6 @@ let conditionalTokensService;
 const SellOrPositions = ({
   account,
   markets,
-  marketResolutionStates,
   positions,
   collateral,
   lmsrState,
@@ -35,7 +31,6 @@ const SellOrPositions = ({
   // Memoize fetching data files
   const loadDataLayer = useCallback(() => {
     async function getRepo() {
-      conditionalTokensRepo = await getConditionalTokensRepo();
       conditionalTokensService = await getConditionalTokensService();
     }
     getRepo();
@@ -73,6 +68,7 @@ const SellOrPositions = ({
         positions,
         positionBalances
       );
+      console.log("position groups", positionGroups);
       setPositionGroups(
         positionGroups
         //   positionGroups.filter(positionGroup => {
@@ -144,6 +140,7 @@ const SellOrPositions = ({
   );
 
   const clearAllPositions = useCallback(() => {
+    console.log("Call clear all positions");
     setCurrentSellingPosition(null);
     setStagedTransactionType(null);
     setStagedTradeAmounts(null);
@@ -169,96 +166,6 @@ const SellOrPositions = ({
     account
   ]);
 
-  const allMarketsResolved =
-    marketResolutionStates &&
-    marketResolutionStates.every(({ isResolved }) => isResolved);
-  const [redemptionAmount, setRedemptionAmount] = useState(null);
-  useEffect(() => {
-    setRedemptionAmount(
-      allMarketsResolved && positionBalances != null
-        ? positionBalances.reduce(
-            (payoutSum, balance, positionIndex) =>
-              payoutSum.add(
-                positions[positionIndex].outcomes.reduce(
-                  (payoutProduct, { marketIndex, outcomeIndex }) =>
-                    payoutProduct
-                      .mul(
-                        marketResolutionStates[marketIndex].payoutNumerators[
-                          outcomeIndex
-                        ]
-                      )
-                      .div(
-                        marketResolutionStates[marketIndex].payoutDenominator
-                      ),
-                  balance
-                )
-              ),
-            toBN(0)
-          )
-        : null
-    );
-  }, [positions, positionBalances, marketResolutionStates]);
-
-  const redeemPositions = useCallback(async () => {
-    if (!allMarketsResolved)
-      throw new Error("Can't redeem until all markets resolved");
-
-    async function redeemPositionsThroughAllMarkets(
-      marketsLeft,
-      parentCollectionId
-    ) {
-      if (marketsLeft === 0) return;
-
-      const market = markets[marketsLeft - 1];
-      const indexSets = [];
-
-      for (
-        let outcomeIndex = 0;
-        outcomeIndex < market.outcomes.length;
-        outcomeIndex++
-      ) {
-        const outcome = market.outcomes[outcomeIndex];
-        const childCollectionId = combineCollectionIds([
-          parentCollectionId,
-          outcome.collectionId
-        ]);
-
-        const childPositionId = getPositionId(
-          collateral.address,
-          childCollectionId
-        );
-
-        await redeemPositionsThroughAllMarkets(
-          marketsLeft - 1,
-          childCollectionId
-        );
-
-        if (
-          (await conditionalTokensRepo.balanceOf(account, childPositionId)).gtn(
-            0
-          )
-        ) {
-          indexSets.push(toBN(1).shln(outcomeIndex));
-        }
-      }
-
-      if (indexSets.length > 0) {
-        await conditionalTokensRepo.redeemPositions(
-          collateral.address,
-          parentCollectionId,
-          market.conditionId,
-          indexSets,
-          account
-        );
-      }
-    }
-
-    await redeemPositionsThroughAllMarkets(
-      markets.length,
-      `0x${"0".repeat(64)}`
-    );
-  }, [collateral, account, conditionalTokensRepo, allMarketsResolved]);
-
   const isSelling = currentSellingPosition != null;
 
   return isSelling ? (
@@ -282,13 +189,9 @@ const SellOrPositions = ({
   ) : (
     <Positions
       positionGroups={positionGroups}
-      allMarketsResolved={allMarketsResolved}
-      redemptionAmount={redemptionAmount}
       collateral={collateral}
-      redeemPositions={redeemPositions}
       setError={setError}
       ongoingTransactionType={ongoingTransactionType}
-      asWrappedTransaction={asWrappedTransaction}
       probabilities={probabilities}
       positionBalances={positionBalances}
       estimatedSaleEarnings={estimatedSaleEarnings}
@@ -299,12 +202,12 @@ const SellOrPositions = ({
 };
 
 SellOrPositions.propTypes = {
+  account: PropTypes.string.isRequired,
   markets: PropTypes.arrayOf(
     PropTypes.shape({
       conditionId: PropTypes.string.isRequired
     }).isRequired
   ).isRequired,
-  marketResolutionStates: PropTypes.array,
   positions: PropTypes.arrayOf(
     PropTypes.shape({
       positionIndex: PropTypes.number.isRequired,
