+++ app/src/Sidebar/components/Sell/positions.scss
@@ -1,5 +1,4 @@
 @import "~scss/_variables.scss";
-// @import "../../CategoricalSidebar/Positions/sidebar.scss";
 
 .positions {
   &.resolved {
